+++ modules/justin_version.py
@@ -1,2 +1,2 @@
 # This file must by valid Python AND valid Bash!
-justinVersion='02.00.rc0'
+justinVersion='02.00.rc1'